+++ frontend/src/features/recommendList/hooks/use-recommend-list.ts
@@ -1,17 +1,18 @@
 import { useAppDispatch, useAppSelector } from "app/store/store"
 import { selectRecommendList } from "../model/recommend-list-selectors"
 import { actionSearchBooks } from "../model/recommend-list-actions"
+import { BooksRequest } from "shared/types/request"
 
 
 export const useRecommendList = (): [
     ReturnType<typeof selectRecommendList>,
-    (search: string) => void
+    (search: BooksRequest) => void
 ] =>
 {
     const dispatch = useAppDispatch()
 
     const booksInfo = useAppSelector(selectRecommendList)
-    const searchBooks = (search: string) => dispatch(actionSearchBooks(search))
+    const searchBooks = (searchRequest: BooksRequest) => dispatch(actionSearchBooks(searchRequest))
 
     return [{...booksInfo}, searchBooks]
 }
